Move tests from DDSketchTest to QuantileSketchTest
