+++ src/test/java/com/datadoghq/sketch/ddsketch/DDSketchTest.java
@@ -54,13 +54,17 @@ abstract class DDSketchTest extends QuantileSketchTest<DDSketch> {
   @Override
   protected void assertQuantileAccurate(
       boolean merged, double[] sortedValues, double quantile, double actualQuantileValue) {
+    assertQuantileAccurate(sortedValues, quantile, actualQuantileValue, relativeAccuracy());
+  }
 
+  static void assertQuantileAccurate(
+      double[] sortedValues, double quantile, double actualQuantileValue, double relativeAccuracy) {
     final double lowerQuantileValue =
         sortedValues[(int) Math.floor(quantile * (sortedValues.length - 1))];
     final double upperQuantileValue =
         sortedValues[(int) Math.ceil(quantile * (sortedValues.length - 1))];
 
-    assertAccurate(lowerQuantileValue, upperQuantileValue, actualQuantileValue);
+    assertAccurate(lowerQuantileValue, upperQuantileValue, actualQuantileValue, relativeAccuracy);
   }
 
   @Override
@@ -90,14 +94,19 @@ abstract class DDSketchTest extends QuantileSketchTest<DDSketch> {
   }
 
   private void assertAccurate(double minExpected, double maxExpected, double actual) {
+    assertAccurate(minExpected, maxExpected, actual, relativeAccuracy());
+  }
+
+  private static void assertAccurate(
+      double minExpected, double maxExpected, double actual, double relativeAccuracy) {
     final double relaxedMinExpected =
         minExpected > 0
-            ? minExpected * (1 - relativeAccuracy())
-            : minExpected * (1 + relativeAccuracy());
+            ? minExpected * (1 - relativeAccuracy)
+            : minExpected * (1 + relativeAccuracy);
     final double relaxedMaxExpected =
         maxExpected > 0
-            ? maxExpected * (1 + relativeAccuracy())
-            : maxExpected * (1 - relativeAccuracy());
+            ? maxExpected * (1 + relativeAccuracy)
+            : maxExpected * (1 - relativeAccuracy);
 
     if (actual < relaxedMinExpected - AccuracyTester.FLOATING_POINT_ACCEPTABLE_ERROR
         || actual > relaxedMaxExpected + AccuracyTester.FLOATING_POINT_ACCEPTABLE_ERROR) {
@@ -107,101 +116,6 @@ abstract class DDSketchTest extends QuantileSketchTest<DDSketch> {
 
   private void assertAccurate(double expected, double actual) {
     assertAccurate(expected, expected, actual);
-  }
-
-  @Test
-  void testNegativeConstants() {
-    testAdding(0);
-    testAdding(-1);
-    testAdding(-1, -1, -1);
-    testAdding(-10, -10, -10);
-    testAdding(IntStream.range(0, 10000).mapToDouble(i -> -2).toArray());
-    testAdding(-10, -10, -11, -11, -11);
-  }
-
-  @Test
-  void testNegativeAndPositiveConstants() {
-    testAdding(0);
-    testAdding(-1, 1);
-    testAdding(-1, -1, -1, 1, 1, 1);
-    testAdding(-10, -10, -10, 10, 10, 10);
-    testAdding(IntStream.range(0, 20000).mapToDouble(i -> i % 2 == 0 ? 2 : -2).toArray());
-    testAdding(-10, -10, -11, -11, -11, 10, 10, 11, 11, 11);
-  }
-
-  @Test
-  void testWithZeros() {
-
-    testAdding(IntStream.range(0, 100).mapToDouble(i -> 0).toArray());
-
-    testAdding(
-        DoubleStream.concat(
-                IntStream.range(0, 10).mapToDouble(i -> 0),
-                IntStream.range(-100, 100).mapToDouble(i -> i))
-            .toArray());
-
-    testAdding(
-        DoubleStream.concat(
-                IntStream.range(-100, 100).mapToDouble(i -> i),
-                IntStream.range(0, 10).mapToDouble(i -> 0))
-            .toArray());
-  }
-
-  @Test
-  void testWithoutZeros() {
-    testAdding(
-        DoubleStream.concat(
-                IntStream.range(-100, -1).mapToDouble(i -> i),
-                IntStream.range(1, 100).mapToDouble(i -> i))
-            .toArray());
-  }
-
-  @Test
-  void testNegativeNumbersIncreasingLinearly() {
-    testAdding(IntStream.range(-10000, 0).mapToDouble(v -> v).toArray());
-  }
-
-  @Test
-  void testNegativeAndPositiveNumbersIncreasingLinearly() {
-    testAdding(IntStream.range(-10000, 10000).mapToDouble(v -> v).toArray());
-  }
-
-  @Test
-  void testNegativeNumbersDecreasingLinearly() {
-    testAdding(IntStream.range(0, 10000).mapToDouble(v -> -v).toArray());
-  }
-
-  @Test
-  void testNegativeAndPositiveNumbersDecreasingLinearly() {
-    testAdding(IntStream.range(0, 20000).mapToDouble(v -> 10000 - v).toArray());
-  }
-
-  @Test
-  void testNegativeNumbersIncreasingExponentially() {
-    testAdding(IntStream.range(0, 100).mapToDouble(i -> -Math.exp(i)).toArray());
-  }
-
-  @Test
-  void testNegativeAndPositiveNumbersIncreasingExponentially() {
-    testAdding(
-        DoubleStream.concat(
-                IntStream.range(0, 100).mapToDouble(i -> -Math.exp(i)),
-                IntStream.range(0, 100).mapToDouble(Math::exp))
-            .toArray());
-  }
-
-  @Test
-  void testNegativeNumbersDecreasingExponentially() {
-    testAdding(IntStream.range(0, 100).mapToDouble(i -> -Math.exp(-i)).toArray());
-  }
-
-  @Test
-  void testNegativeAndPositiveNumbersDecreasingExponentially() {
-    testAdding(
-        DoubleStream.concat(
-                IntStream.range(0, 100).mapToDouble(i -> -Math.exp(-i)),
-                IntStream.range(0, 100).mapToDouble(i -> Math.exp(-i)))
-            .toArray());
   }
 
   @Override
